requirements.txt
boto3>=1.28.0
python-dateutil>=2.8.2
streamlit>=1.24.0
python-dotenv>=0.19.0
Pillow>=10.0.0
PyPDF2>=3.0.0
